+++ frontend/src/routes/(app)/dashboard/[sides]/+layout.server.ts
