+++ frontend/src/routes/(app)/dashboard/[sides]/+page.svelte
