+++ frontend/src/routes/(app)/dashboard/[sides]/configuration/+page.svelte
