+++ frontend/src/routes/(app)/dashboard/[sides]/palettes/palette-editor.svelte
