+++ frontend/src/routes/(app)/dashboard/[sides]/playerskins/+page.svelte
