+++ frontend/src/routes/(app)/dashboard/[sides]/playerskins/player-skin-editor.svelte
